[ao] Update DynamicStatic Detector to account for Conv
Summary: This updates the DynamicStatic Detector to also provide insight
into whether Conv layers should use dynamic or static quantization.
Before, this was not included because as of now, Dynamic quantization is
not supported for Conv layers, but this adds a check for Conv layers and
if dynamic is recommended, it will also give a disclaimer that it is not
currently supported but will be in the future.

Test Plan: python test/test_quantization.py TestFxModelReportDetectDynamicStatic

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 5237c28f1e6f0ee57f6f643c662ef83e748bacf4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81972